Fixed potential error with instance variables
When the controller is in standby, some instance variables like ramv and
ramw would have retained their last values. As the get velocity methods
call upon these variables, there is the potential for the controller to
output nonzero velocities and hence drive the robot around in standby.
Fix is to zero ALL instance variables in standby (which is redundant but
whatever :P